model_cvs_docs/requirements.txt
fastapi
uvicorn
pillow
pdfplumber
python-docx
pytesseract
python-multipart
rapidfuzz

huggingface_hub
